academic-paper-qa/requirements.txt
# ============================================================================
# LlamaIndex 核心和集成包
# ============================================================================

# LlamaIndex 核心
llama-index>=0.13.0,<0.14.0
llama-index-core>=0.13.0,<0.14.0

# LlamaIndex LLM 集成（仅保留实际使用的）
llama-index-llms-openai>=0.3.0         # OpenAI 及兼容 API（Moonshot、DeepSeek 等）

# LlamaIndex Embedding 集成
llama-index-embeddings-huggingface>=0.3.0  # bge-small-zh-v1.5 等本地模型
llama-index-embeddings-openai>=0.3.0       # OpenAI Embedding（可选）

# LlamaIndex 向量数据库集成
llama-index-vector-stores-chroma>=0.3.0    # Chroma 向量数据库（主要使用）
# llama-index-vector-stores-qdrant>=0.3.0  # Qdrant（可选，未使用可注释）

# LlamaIndex 文档加载器集成
# llama-index-readers-file>=0.3.0          # 可选，项目使用自定义加载器

# ============================================================================
# 文档处理（仅保留必需的）
# ============================================================================

# PDF 处理（SimpleDirectoryReader 已支持基础 PDF）
# pypdf>=4.0.0                              # 基础 PDF 支持（可选）
# pdfplumber>=0.10.0                        # 高级 PDF 解析（可选）
# pymupdf>=1.23.0                           # 更强大的 PDF 解析（可选）

# DOCX 处理（SimpleDirectoryReader 已支持）
# python-docx>=1.1.0                        # 可选

# Markdown 处理（Python 内置支持）
# markdown>=3.5.0                           # 可选

# 通用文档处理（较重，仅在需要复杂文档解析时安装）
# unstructured>=0.12.0                      # 强大但依赖多，可选
# python-magic>=0.4.27                      # 文件类型检测（可选）

# ============================================================================
# 向量数据库客户端
# ============================================================================

chromadb>=0.4.22                           # Chroma 客户端（必需）
# qdrant-client>=1.7.0                     # Qdrant 客户端（可选）

# ============================================================================
# 模型推理（本地 Embedding 模型）
# ============================================================================

# Transformers 生态（用于本地 Embedding）
sentence-transformers>=2.5.0               # 本地 Embedding 模型（必需）
transformers>=4.40.0                       # Transformers 核心
torch>=2.1.0                               # PyTorch（sentence-transformers 依赖）

# 加速推理（可选）
# accelerate>=0.27.0                       # GPU 加速（有 GPU 时安装）

# ============================================================================
# Web 搜索工具
# ============================================================================

duckduckgo-search>=4.0.0                   # DuckDuckGo 搜索（必需）
requests>=2.31.0                           # HTTP 请求（必需）
urllib3<2.0                                # 版本限制，避免与 sentence-transformers 冲突
# beautifulsoup4>=4.12.0                   # HTML 解析（项目未使用）
# httpx>=0.26.0                            # 异步 HTTP（项目未使用）

# ============================================================================
# Web UI 界面
# ============================================================================

gradio>=4.0.0                              # Web UI 框架（必需）

# ============================================================================
# 环境配置和工具
# ============================================================================

python-dotenv>=1.0.0                       # 环境变量管理（必需）
# pydantic>=2.5.0                          # 数据验证（LlamaIndex 已包含）
# pydantic-settings>=2.1.0                 # Settings 管理（LlamaIndex 已包含）

# ============================================================================
# 日志和监控
# ============================================================================

loguru>=0.7.0                              # 日志库（必需）
# tenacity>=8.2.0                          # 重试机制（可选，项目未直接使用）

# ============================================================================
# 开发工具（开发环境安装，生产环境可注释）
# ============================================================================

# 代码质量（可选）
# black>=24.0.0
# isort>=5.13.0
# flake8>=7.0.0
# mypy>=1.8.0

# 测试（可选）
pytest>=8.0.0
# pytest-asyncio>=0.23.0
# pytest-cov>=4.1.0

# Jupyter 支持（可选，用于开发和演示）
# jupyter>=1.0.0
# notebook>=7.0.0
# ipywidgets>=8.1.0
